fix: moves component to Sections
